Extract timestamp as double for InMemorySessionService events
InMemorySessionService sets a Session's last modified time based on when
the last appended event's timestamp. The timestamp in an event is
recorded in millis while the Session's timestamp is an Instant. During
the transformation, Events perform this converstion using division.
Before this change, the timestamp was truncated to the second, yet the
code was trying to extract nanos which were always 0. This fixes that
bug with a simple type change. I've also added a test to prevent
regressions.

+++ core/src/main/java/com/google/adk/sessions/InMemorySessionService.java
@@ -163,8 +163,8 @@ public final class InMemorySessionService implements BaseSessionService {
   }
 
   // Helper to get event timestamp as epoch seconds
-  private long getEventTimestampEpochSeconds(Event event) {
-    return event.timestamp() / 1000L;
+  private double getEventTimestampEpochSeconds(Event event) {
+    return event.timestamp() / 1000.0;
   }
 
   @Override
